LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/scroll-timeline-axis-writing-mode-expected.txt


FAIL Initial axis assert_equals: expected "175px" but got "100.125px"
FAIL Vertical axis assert_equals: expected "175px" but got "100.453125px"
FAIL Horizontal axis assert_equals: expected "125px" but got "100.796875px"
FAIL Block axis in horizontal writing-mode assert_equals: expected "175px" but got "101.125px"
FAIL Inline axis in horizontal writing-mode assert_equals: expected "125px" but got "101.453125px"
FAIL Block axis in vertical writing-mode assert_equals: expected "125px" but got "101.78125px"
FAIL Inline axis in vertical writing-mode assert_equals: expected "175px" but got "102.109375px"
